Used NextStatementIndexFetch in resolve_labels.
commit-id:b010986e
