+++ crates/sierra_generator/src/resolve_labels.rs
@@ -2,10 +2,9 @@
 #[path = "resolve_labels_test.rs"]
 mod test;
 
-use std::collections::{hash_map, HashMap};
-
 use sierra::program;
 
+use crate::next_statement_index_fetch::NextStatementIndexFetch;
 use crate::pre_sierra;
 
 /// Replaces labels with their corresponding StatementIdx.
@@ -27,43 +26,13 @@ pub fn resolve_labels(
         .collect()
 }
 
-/// Returns a map from LabelId to the index of the next Sierra statement.
-fn get_label_id_to_index(
-    statements: &[pre_sierra::Statement],
-) -> HashMap<pre_sierra::LabelId, usize> {
-    let mut label_id_to_index = HashMap::<pre_sierra::LabelId, usize>::new();
-    let mut index = 0;
-    for statement in statements {
-        match &statement {
-            pre_sierra::Statement::Sierra(_) => {
-                index += 1;
-            }
-            pre_sierra::Statement::Label(label) => {
-                if let hash_map::Entry::Vacant(entry) = label_id_to_index.entry(label.id) {
-                    entry.insert(index);
-                } else {
-                    panic!("Label {} was already declared.", label.id)
-                }
-            }
-            pre_sierra::Statement::PushValues(_) => {
-                panic!("Unexpected pre_sierra::Statement::PushValues in get_label_id_to_index().")
-            }
-        }
-    }
-    label_id_to_index
-}
-
 /// Helper struct for resolve_labels.
 pub struct LabelReplacer {
-    label_id_to_index: HashMap<pre_sierra::LabelId, usize>,
+    next_statement_index_fetch: NextStatementIndexFetch,
 }
 impl LabelReplacer {
-    fn new(label_id_to_index: HashMap<pre_sierra::LabelId, usize>) -> Self {
-        Self { label_id_to_index }
-    }
-
     pub fn from_statements(statements: &[pre_sierra::Statement]) -> LabelReplacer {
-        Self::new(get_label_id_to_index(statements))
+        Self { next_statement_index_fetch: NextStatementIndexFetch::new(statements, false) }
     }
 
     /// Replaces the pre-sierra labels in the given statement, and returns [program::Statement].
@@ -122,7 +91,6 @@ impl LabelReplacer {
 
     /// Resolves the given pre-sierra label, and returns [program::StatementIdx].
     pub fn handle_label_id(&self, label_id: pre_sierra::LabelId) -> program::StatementIdx {
-        // TODO(lior): handle missing labels.
-        program::StatementIdx(self.label_id_to_index[&label_id])
+        program::StatementIdx(self.next_statement_index_fetch.resolve_label(&label_id))
     }
 }
